requirements.txt
# Основные зависимости
sentence-transformers==2.3.1      # Обновлено для совместимости
faiss-cpu==1.7.4
llama-index-core==0.12.1
llama-index-vector-stores-faiss==0.12.1
llama-index-embeddings-openai==0.1.7
python-dotenv==1.0.0
watchdog==3.0.0
requests==2.31.0
numpy==1.26.4
huggingface-hub==0.23.3

# Тестирование
pytest==7.4.0
pytest-mock==3.14.0

# Дополнительные
streamlit==1.28.0
python-json-logger==2.0.7
tqdm==4.66.1